extracted inventory_binds_handler and pick_bind_handler
